rename the multiclass folder to multiclass_extension
