+++ qiskit_aqua/algorithms/components/multiclass_extension/estimator.py
