+++ qiskit_aqua/algorithms/components/multiclass_extension/multiclass_extension.py
